Rename: Rename to InspireTale Utils
