+++ Assets/Plugins/InspireTale.meta
