+++ Assets/Plugins/InspireTale/Prefabs/SceneFadingTransition.cs.meta
