+++ Assets/Plugins/InspireTale/Scripts.meta
