+++ Assets/Plugins/InspireTale/Scripts/Components.meta
