+++ Assets/Plugins/InspireTale/Scripts/Components/TextAdjuster.cs
@@ -1,40 +1,50 @@
 ﻿//  *******STANDARD IMPORT*******
 using UnityEngine;
-using TMPro;
+using UnityEngine.UI;
 
 //  *******LOCAL IMPORT*******
 
-namespace InspireTaleFramework
+namespace InspireTale.Utils
 {
     /// <summary>This component will be adjust Thai unicode string
     /// for display in Unity
-    /// - Require "TextMeshProGUI" class to be attach on object
+    /// - Require "Text" class to be attach on object
     /// - To get and set text you can use "text" property
     ///     <example>textAdjuster.text = "hello"<example>
-    /// - To access "TextMeshProGUI" component you can access by "component" property
+    /// - To access "Text" component you can access by "component" property
     ///     <example>textAdjuster.component.AddEventListener(...)</example>
     ///</summary>
-    public class TextMeshAdjuster : MonoBehaviour
+    public class TextAdjuster: MonoBehaviour
     {
         //  *******PROPERTY*******
-        //  property to get and set TextMeshProRUI string
+        //  property to get and set Text string
         public string text
         {
+            //  when get text return unadjest string
             get { return unAdjustString; }
+
+            //  when set text will
+            //  1. set unadjustString as input value
+            //  2. set adjustString with ThaiFontAdjuster with unadjustString
             set {
                 this.unAdjustString = value;
                 this.adjustedString = ThaiFontAdjuster.Adjust(this.unAdjustString);
 
-                if (this.m_textMeshProGUI == null)
+                if (this.m_text == null)
                 {
-                    this.m_textMeshProGUI = GetComponent<TextMeshProUGUI>();
+                    this.m_text = GetComponent<Text>();   
                 }
 
-                this.m_textMeshProGUI.text = this.text;
+                //  update value to Text component
+                this.m_text.text = this.text;
             }
         }
 
-        //  shortcut to access TextMeshProRUI component
+        //  shortcut to access Text component
+        public Text component
+        {
+            get { return this.m_text; }
+        }
 
         //  *******PUBLIC VARIABLE*******
 
@@ -45,13 +55,14 @@ namespace InspireTaleFramework
         //  display value of text string
         private string unAdjustString;
 
-        private TextMeshProUGUI m_textMeshProGUI;
+        //  Text component instance
+        private Text m_text;
 
         //  *******UNITY HOOK*******
-        void Awake()
+        void Awake ()
         {
-            this.m_textMeshProGUI = GetComponent<TextMeshProUGUI>();
-            this.text = this.m_textMeshProGUI.text;
+            this.m_text = GetComponent<Text>();
+            this.text = this.m_text.text;
         }
 
         //  *******PUBLIC FUNCTION*******
@@ -59,4 +70,3 @@ namespace InspireTaleFramework
         //  *******PRIVATE FUNCTION*******
     }
 }
-
