+++ Assets/Plugins/InspireTale/Scripts/Components/TextAdjuster.cs.meta
