+++ Assets/Plugins/InspireTale/Scripts/Components/TextMeshAdjuster.cs
@@ -1,50 +1,40 @@
 ﻿//  *******STANDARD IMPORT*******
 using UnityEngine;
-using UnityEngine.UI;
+using TMPro;
 
 //  *******LOCAL IMPORT*******
 
-namespace InspireTaleFramework
+namespace InspireTale.Utils
 {
     /// <summary>This component will be adjust Thai unicode string
     /// for display in Unity
-    /// - Require "Text" class to be attach on object
+    /// - Require "TextMeshProGUI" class to be attach on object
     /// - To get and set text you can use "text" property
     ///     <example>textAdjuster.text = "hello"<example>
-    /// - To access "Text" component you can access by "component" property
+    /// - To access "TextMeshProGUI" component you can access by "component" property
     ///     <example>textAdjuster.component.AddEventListener(...)</example>
     ///</summary>
-    public class TextAdjuster: MonoBehaviour
+    public class TextMeshAdjuster : MonoBehaviour
     {
         //  *******PROPERTY*******
-        //  property to get and set Text string
+        //  property to get and set TextMeshProRUI string
         public string text
         {
-            //  when get text return unadjest string
             get { return unAdjustString; }
-
-            //  when set text will
-            //  1. set unadjustString as input value
-            //  2. set adjustString with ThaiFontAdjuster with unadjustString
             set {
                 this.unAdjustString = value;
                 this.adjustedString = ThaiFontAdjuster.Adjust(this.unAdjustString);
 
-                if (this.m_text == null)
+                if (this.m_textMeshProGUI == null)
                 {
-                    this.m_text = GetComponent<Text>();   
+                    this.m_textMeshProGUI = GetComponent<TextMeshProUGUI>();
                 }
 
-                //  update value to Text component
-                this.m_text.text = this.text;
+                this.m_textMeshProGUI.text = this.text;
             }
         }
 
-        //  shortcut to access Text component
-        public Text component
-        {
-            get { return this.m_text; }
-        }
+        //  shortcut to access TextMeshProRUI component
 
         //  *******PUBLIC VARIABLE*******
 
@@ -55,14 +45,13 @@ namespace InspireTaleFramework
         //  display value of text string
         private string unAdjustString;
 
-        //  Text component instance
-        private Text m_text;
+        private TextMeshProUGUI m_textMeshProGUI;
 
         //  *******UNITY HOOK*******
-        void Awake ()
+        void Awake()
         {
-            this.m_text = GetComponent<Text>();
-            this.text = this.m_text.text;
+            this.m_textMeshProGUI = GetComponent<TextMeshProUGUI>();
+            this.text = this.m_textMeshProGUI.text;
         }
 
         //  *******PUBLIC FUNCTION*******
@@ -70,3 +59,4 @@ namespace InspireTaleFramework
         //  *******PRIVATE FUNCTION*******
     }
 }
+
