+++ Assets/Plugins/InspireTale/Scripts/Components/TextMeshAdjuster.cs.meta
