+++ Assets/Plugins/InspireTale/Scripts/Enums.meta
