+++ Assets/Plugins/InspireTale/Scripts/Enums/MouseButton.cs.meta
