+++ Assets/Plugins/InspireTale/Scripts/MonoSingleton.cs.meta
