+++ Assets/Plugins/InspireTale/Scripts/Singleton.cs
@@ -1,4 +1,4 @@
-﻿namespace InspireTaleFramework
+﻿namespace InspireTale.Utils
 {
     public class Singleton<T> where T : new()
     {
