+++ Assets/Plugins/InspireTale/Scripts/Singleton.cs.meta
