+++ Assets/Plugins/InspireTale/Scripts/Util.meta
