+++ Assets/Plugins/InspireTale/Scripts/Util/RandomUtil.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine;
 
-namespace InspireTaleFramework
+namespace InspireTale.Utils
 {
     public class RandomUtil
     {
