+++ Assets/Plugins/InspireTale/Scripts/Util/ScreenUtil.cs.meta
